requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
torch
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
pysdf
dearpygui
packaging
scipy
lpips
imageio
clip
imageio_ffmpeg
scikit-spatial